Change casting from float to double when using printf.

+++ trunk/src/PerplexityOptimizer.cpp
@@ -128,7 +128,7 @@ PerplexityOptimizer::Optimize(ParamVector &params, Optimization technique) {
 
     Logger::Log(1, "Iterations    = %i\n", numIter);
     Logger::Log(1, "Elapsed Time  = %f\n",
-                float(endTime - startTime) / CLOCKS_PER_SEC);
+                (double)(endTime - startTime) / CLOCKS_PER_SEC);
     Logger::Log(1, "Perplexity    = %f\n", exp(minEntropy));
     Logger::Log(1, "Num OOVs      = %lu\n", _numOOV);
     Logger::Log(1, "Num ZeroProbs = %lu\n", _numZeroProbs);
