+++ trunk/src/util/Logger.cpp
@@ -52,7 +52,7 @@ void Logger::Log(int level, const char *fmt, ...) {
         va_list args;
         va_start(args, fmt);
         if (_timestamp) 
-            printf("%.3f\t", (float)(clock() - _startTime) / CLOCKS_PER_SEC);
+            printf("%.3f\t", (double)(clock() - _startTime) / CLOCKS_PER_SEC);
         vprintf(fmt, args);
         va_end(args);
     }
@@ -64,7 +64,7 @@ void Logger::Warn(int level, const char *fmt, ...) {
         va_start(args, fmt);
         fprintf(stderr, "\e[0;33m");
         if (_timestamp) 
-            printf("%.3f\t", (float)(clock() - _startTime) / CLOCKS_PER_SEC);
+            printf("%.3f\t", (double)(clock() - _startTime) / CLOCKS_PER_SEC);
         vfprintf(stderr, fmt, args);
         fprintf(stderr, "\e[m");
         va_end(args);
@@ -77,7 +77,7 @@ void Logger::Error(int level, const char *fmt, ...) {
         va_start(args, fmt);
         fprintf(stderr, "\e[1;31m");
         if (_timestamp) 
-            printf("%.3f\t", (float)(clock() - _startTime) / CLOCKS_PER_SEC);
+            printf("%.3f\t", (double)(clock() - _startTime) / CLOCKS_PER_SEC);
         vfprintf(stderr, fmt, args);
         fprintf(stderr, "\e[m");
         va_end(args);
